+++ app/Console/Commands/DemandProcessPayments.php
@@ -21,16 +21,20 @@ declare(strict_types = 1);
 
 namespace Adshares\Adserver\Console\Commands;
 
+use Adshares\Adserver\Exceptions\ConsoleCommandException;
+use Adshares\Adserver\Models\Config;
 use Adshares\Adserver\Models\PaymentReport;
 use Adshares\Common\Exception\InvalidArgumentException;
 use DateTime;
+use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Support\Facades\Artisan;
 use Symfony\Component\Console\Exception\LogicException;
 
 class DemandProcessPayments extends BaseCommand
 {
     protected $signature = 'ops:demand:payments:process
-                            {--f|from= : Date from which reports will be processed}';
+                            {--f|from= : Date from which reports will be processed}
+                            {--ids= : Report ids to process}';
 
     protected $description = 'Fetches reports, sends payments and updates statistics';
 
@@ -45,18 +49,28 @@ class DemandProcessPayments extends BaseCommand
         $this->info('Start command '.$this->getName());
 
         PaymentReport::fillMissingReports();
-        $reports = PaymentReport::fetchUndone($this->getReportDateFrom());
+        $lastAvailableTimestamp = $this->getLastAvailableTimestamp();
+        $isRecalculationNeeded = null !== $this->option('ids');
+        $reports = $this->fetchPaymentReports();
 
         /** @var PaymentReport $report */
         foreach ($reports as $report) {
             $timestamp = $report->id;
+            if ($timestamp > $lastAvailableTimestamp) {
+                continue;
+            }
 
             if ($report->isNew()) {
+                $getPaymentsParameters = ['--timestamp' => $timestamp];
+                if ($isRecalculationNeeded) {
+                    $getPaymentsParameters['--recalculate'] = 1;
+                }
+
                 try {
                     $status =
                         Artisan::call(
                             AdPayGetPayments::COMMAND_SIGNATURE,
-                            ['--timestamp' => $timestamp],
+                            $getPaymentsParameters,
                             $this->getOutput()
                         );
                 } catch (LogicException $logicException) {
@@ -76,12 +90,17 @@ class DemandProcessPayments extends BaseCommand
             }
 
             if ($report->isUpdated()) {
-                $parameters = [
+                $preparePaymentsParameters = [
                     '--from' => (new DateTime('@'.$timestamp))->format(DateTime::ATOM),
                     '--to' => (new DateTime('@'.($timestamp + 3599)))->format(DateTime::ATOM),
                 ];
+
                 try {
-                    Artisan::call(DemandPreparePayments::COMMAND_SIGNATURE, $parameters, $this->getOutput());
+                    Artisan::call(
+                        DemandPreparePayments::COMMAND_SIGNATURE,
+                        $preparePaymentsParameters,
+                        $this->getOutput()
+                    );
                 } catch (LogicException $logicException) {
                     $this->warn(sprintf('Command %s is locked', DemandPreparePayments::COMMAND_SIGNATURE));
 
@@ -123,6 +142,25 @@ class DemandProcessPayments extends BaseCommand
         $this->info('End command '.$this->getName());
     }
 
+    private function fetchPaymentReports(): Collection
+    {
+        if (null !== ($ids = $this->option('ids'))) {
+            $reports = PaymentReport::fetchByIds(explode(',', $ids));
+
+            if ($reports->isEmpty()) {
+                throw new ConsoleCommandException(sprintf('Payment reports "%s" not found', $ids));
+            }
+
+            return $reports->each(
+                function (PaymentReport $report) {
+                    $report->setNew();
+                }
+            );
+        }
+
+        return PaymentReport::fetchUndone($this->getReportDateFrom());
+    }
+
     private function getReportDateFrom(): DateTime
     {
         $optionFrom = $this->option('from');
@@ -138,5 +176,18 @@ class DemandProcessPayments extends BaseCommand
         }
 
         return new DateTime('@'.$timestampFrom);
+    }
+
+    private function getLastAvailableTimestamp(): int
+    {
+        return $this->getLastExportedEventTimestamp() - PaymentReport::MIN_INTERVAL;
+    }
+
+    private function getLastExportedEventTimestamp(): int
+    {
+        return min(
+            Config::fetchDateTime(Config::ADPAY_LAST_EXPORTED_EVENT_TIME)->getTimestamp(),
+            Config::fetchDateTime(Config::ADPAY_LAST_EXPORTED_CONVERSION_TIME)->getTimestamp()
+        );
     }
 }